Small modifications for Android compilation

+++ include/xgboost/c_api.h
@@ -24,7 +24,7 @@ XGB_EXTERN_C {
 #endif
 
 // manually define unsign long
-typedef unsigned long bst_ulong;  // NOLINT(*)
+typedef unsigned long long bst_ulong;  // NOLINT(*)
 
 /*! \brief handle to DMatrix */
 typedef void *DMatrixHandle;
@@ -40,7 +40,7 @@ typedef struct {
   /*! \brief number of rows in the minibatch */
   size_t size;
   /*! \brief row pointer to the rows in the data */
-  long* offset;  // NOLINT(*)
+  long long* offset;  // NOLINT(*)
   /*! \brief labels of each instance */
   float* label;
   /*! \brief weight of each instance, can be NULL */
